+++ modules/hanna-react/src/_abstract/_TogglerGroup.tsx
@@ -1,6 +1,6 @@
 import React, { useMemo } from 'react';
-import { useDomid } from '@hugsmidjan/react/hooks';
 import { modifiedClass } from '@hugsmidjan/qj/classUtils';
+import { useDomid } from '@hugsmidjan/react/hooks';
 
 import { FormFieldInputProps } from '../FormField.js';
 import { HTMLProps, useMixedControlState } from '../utils.js';