+++ modules/hanna-react/src/_abstract/_TogglerInput.tsx
@@ -2,6 +2,7 @@ import React, { Fragment } from 'react';
 import { modifiedClass } from '@hugsmidjan/qj/classUtils';
 import { useDomid } from '@hugsmidjan/react/hooks';
 
+import { WrapperElmProps } from '../utils.js';
 import { BemModifierProps } from '../utils/types.js';
 
 export type TogglerInputProps = {
@@ -18,9 +19,9 @@ export type TogglerInputProps = {
   reqText?: string | false;
   errorMessage?: string | JSX.Element;
   Wrapper?: 'div' | 'li';
-  wrapperProps?: JSX.IntrinsicElements['div'];
   inputProps?: JSX.IntrinsicElements['input'];
 } & BemModifierProps &
+  WrapperElmProps &
   Omit<JSX.IntrinsicElements['input'], 'type'>;
 
 type _TogglerInputProps = {
@@ -54,7 +55,7 @@ export const TogglerInput = (props: TogglerInputProps & _TogglerInputProps) => {
   const reqStar = required && reqText !== false && (
     <abbr
       className={bem + '__label__reqstar'}
-      // TODO: add mo-better i18n thinking
+      // FIXME: add mo-better i18n thinking
       title={(reqText || 'Þarf að haka í') + ': '}
     >
       *
@@ -73,7 +74,12 @@ export const TogglerInput = (props: TogglerInputProps & _TogglerInputProps) => {
   return (
     <Wrapper
       {...(wrapperProps as {})}
-      className={modifiedClass(bem, modifier, className)}
+      className={modifiedClass(
+        bem,
+        modifier,
+        // Prefer `className` over `wrapperProps.className`
+        className || (wrapperProps || {}).className
+      )}
     >
       <input
         className={bem + '__input'}